+++ library/tests/unit/policies/xr0/test_dit.py
@@ -1,15 +1,7 @@
 # Copyright (C) 2026 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-"""Unit tests for the XR0 DiT action head (``qwen3vl_dit``).
-
-Fast, self-contained tests with no external dependencies (no HuggingFace model
-downloads). Small module configs exercise the tensor helpers and DiT decoder
-stack for structural invariants and pinned reference outputs. Weights and inputs
-are drawn from a fixed seed, so a chosen slice of each forward output is
-reproducible and guards against silent behavioural regressions. Exact numerical
-parity against the source implementation is proven separately by the golden
-fixtures (``G10``-``G15``).
+"""Unit tests for the XR0 DiT action head and rectified-flow model (``dit``).
 """
 
 from __future__ import annotations
@@ -17,13 +9,14 @@ from __future__ import annotations
 import pytest
 import torch
 
-from physicalai.policies.xr0.qwen3vl_dit import (
+from physicalai.policies.xr0.dit import (
     DiT,
     DecoderLayer,
     DiTAttention,
     DiTMLP,
     MLPProjector,
     TimestepEmbedder,
+    XR0FlowModel,
     apply_rotary_pos_emb,
     modulate,
     repeat_kv,
@@ -112,19 +105,6 @@ class TestHelpers:
 
 class TestMLPProjector:
     """Tests for MLPProjector."""
-
-    def test_single_layer_is_linear(self) -> None:
-        """num_layers=1, bias=False is a bare linear projection."""
-        proj = MLPProjector(8, 4, num_layers=1, bias=False)
-        assert len(proj.layers) == 1
-        x = torch.randn(3, 8)
-        torch.testing.assert_close(proj(x), x @ proj.layers[0].weight.t())
-
-    def test_bias_flag(self) -> None:
-        """bias=True adds bias to the linear layers."""
-        assert MLPProjector(8, 4, num_layers=1, bias=True).layers[0].bias is not None
-        assert MLPProjector(8, 4, num_layers=1, bias=False).layers[0].bias is None
-
     @pytest.mark.parametrize(
         "reference",
         [torch.tensor([0.06443175, -0.08235656, -0.01462553, -0.19436540, -0.07179737])],
@@ -227,18 +207,6 @@ class TestDecoderLayer:
 
 class TestDiT:
     """Tests for the DiT decoder stack."""
-
-    def test_tail_alignment_with_longer_cache(self) -> None:
-        """A KV-cache longer than layer_num is consumed from its tail."""
-        dit = DiT(hidden_size=HIDDEN, layer_num=2, head_dim=HEAD_DIM, kv_heads=KV_HEADS)
-        assert len(dit.layers) == 2
-        hidden, kv, pos, mask = _attn_inputs()
-        # 4-entry cache with only 2 DiT layers -> start_idx = 2 (uses last two).
-        past_key_values = [(kv[0].clone(), kv[1].clone()) for _ in range(4)]
-        t_embeds = torch.randn(BATCH, 6, HIDDEN)
-        out = dit(hidden, past_key_values, mask, pos, t_embeds)
-        assert out.shape == (BATCH, SEQ, HIDDEN)
-
     @pytest.mark.parametrize(
         "reference",
         [torch.tensor([1.01393712, 0.45370287, -0.99828362, -0.52321780, -0.47535068])],
@@ -250,5 +218,173 @@ class TestDiT:
         hidden, kv, pos, mask = _attn_inputs()
         past_key_values = [kv, (kv[0].clone(), kv[1].clone())]
         t_embeds = torch.randn(BATCH, 6, HIDDEN)
-        out = dit(hidden, past_key_values, mask, pos, t_embeds)[0, 0, :5]
+        out = dit(hidden, past_key_values, mask, pos, t_embeds)
+        assert out.shape == (BATCH, SEQ, HIDDEN)
+        assert torch.allclose(out[0, 0, :5], reference, **TOL), out.tolist()
+
+
+# ============================================================================ #
+# Rectified-flow action model (XR0FlowModel)                                   #
+# ============================================================================ #
+
+# Structural config for the XR0FlowModel tests.
+FLOW_HIDDEN = 128
+FLOW_HEAD_DIM = 128  # DiT default; hidden must be divisible by it.
+FLOW_KV_HEADS = 1
+LAYERS = 2
+STATE_LEN = 1
+ACTION_LEN = 4
+STATE_DIM = 8
+ACTION_DIM = 8
+NUM_STEPS = 3
+
+
+def _build_flow_model() -> XR0FlowModel:
+    """Build a small fp32 flow model matching the shared config."""
+    return XR0FlowModel(
+        state_shape=(STATE_LEN, STATE_DIM),
+        action_shape=(ACTION_LEN, ACTION_DIM),
+        dit_num_layers=LAYERS,
+        dit_hidden_size=FLOW_HIDDEN,
+        dit_kv_heads=FLOW_KV_HEADS,
+        num_steps=NUM_STEPS,
+        dtype=torch.float32,
+    )
+
+
+def _flow_dit_inputs() -> dict:
+    """Build small inputs for ``dit_forward`` / ``_flow_generate``."""
+    q_len = 1 + STATE_LEN + ACTION_LEN  # sink + state + action
+    ang = torch.randn(BATCH, q_len, FLOW_HEAD_DIM)
+    q_causal = torch.tril(torch.ones(q_len, q_len, dtype=torch.bool))
+    cache_ones = torch.ones(q_len, CACHE, dtype=torch.bool)
+    mask = torch.cat([cache_ones, q_causal], dim=-1)[None, None].expand(BATCH, 1, q_len, CACHE + q_len)
+    return {
+        "noisy_action": torch.randn(BATCH, ACTION_LEN, ACTION_DIM),
+        "t": torch.ones(BATCH, 1, 1) * 0.3,
+        "action_mask": torch.ones(BATCH, ACTION_LEN, ACTION_DIM),
+        "state_embed": torch.randn(BATCH, STATE_LEN, FLOW_HIDDEN),
+        "cos": torch.cos(ang),
+        "sin": torch.sin(ang),
+        "past_key_values": [
+            (
+                torch.randn(BATCH, FLOW_KV_HEADS, CACHE, FLOW_HEAD_DIM),
+                torch.randn(BATCH, FLOW_KV_HEADS, CACHE, FLOW_HEAD_DIM),
+            )
+            for _ in range(LAYERS)
+        ],
+        "attn_mask": mask.contiguous(),
+    }
+
+
+def _flow_dit_kwargs(i: dict) -> dict:
+    """Assemble the ``dit_forward`` keyword args from an input bundle."""
+    return {
+        "action_mask": i["action_mask"],
+        "state_embed": i["state_embed"],
+        "position_embeds": (i["cos"], i["sin"]),
+        "past_key_values": i["past_key_values"],
+        "attn_mask": i["attn_mask"],
+        "prefix_length": 0,
+    }
+
+
+class TestFlowMath:
+    """Tests for the pure rectified-flow helpers."""
+
+    def test_interpolate_endpoints(self) -> None:
+        """t=0 returns x0 and t=1 returns x1."""
+        model = _build_flow_model()
+        x0 = torch.randn(BATCH, ACTION_LEN, ACTION_DIM)
+        x1 = torch.randn(BATCH, ACTION_LEN, ACTION_DIM)
+        torch.testing.assert_close(model._flow_interpolate(x0, x1, torch.zeros(BATCH, 1, 1)), x0)
+        torch.testing.assert_close(model._flow_interpolate(x0, x1, torch.ones(BATCH, 1, 1)), x1)
+
+    @pytest.mark.parametrize(
+        "reference",
+        [torch.tensor([0.39629608, -0.72900736, -0.29863209, -0.36008668, 0.27356121])],
+    )
+    def test_interpolate_formula(self, reference: torch.Tensor) -> None:
+        """A seeded interpolation pins a slice of z_t = (1 - t)*x0 + t*x1."""
+        torch.manual_seed(0)
+        model = _build_flow_model()
+        x0 = torch.randn(BATCH, ACTION_LEN, ACTION_DIM)
+        x1 = torch.randn(BATCH, ACTION_LEN, ACTION_DIM)
+        t = torch.rand(BATCH, 1, 1)
+        out = model._flow_interpolate(x0, x1, t)[0, 0, :5]
+        assert torch.allclose(out, reference, **TOL), out.tolist()
+
+    @pytest.mark.parametrize(
+        "reference",
+        [torch.tensor([0.48739055, -0.00960857, -0.06220679, -1.28131175, 2.19818282])],
+    )
+    def test_velocity_target(self, reference: torch.Tensor) -> None:
+        """A seeded velocity target pins a slice of v = x1 - x0."""
+        torch.manual_seed(0)
+        model = _build_flow_model()
+        x0 = torch.randn(BATCH, ACTION_LEN, ACTION_DIM)
+        x1 = torch.randn(BATCH, ACTION_LEN, ACTION_DIM)
+        out = model._flow_velocity_target(x0, x1)[0, 0, :5]
+        assert torch.allclose(out, reference, **TOL), out.tolist()
+
+    def test_sample_timestep_beta_range(self) -> None:
+        """Beta sampling yields shape (batch,) values in (0, 0.999]."""
+        t = _build_flow_model()._sample_timestep(16, dtype=torch.float32)
+        assert t.shape == (16,)
+        assert torch.all(t > 0) and torch.all(t <= 0.999)
+
+    def test_sample_timestep_uniform_fallback(self) -> None:
+        """A non-beta/logit sampler falls back to Uniform(0, 1)."""
+        model = _build_flow_model()
+        model.flow_sampling = "uniform"
+        t = model._sample_timestep(16, dtype=torch.float32)
+        assert torch.all(t >= 0) and torch.all(t < 1)
+
+
+class TestDitForward:
+    """Tests for a single DiT velocity prediction."""
+
+    def test_output_shape(self) -> None:
+        """dit_forward returns (B, action_len, action_dim)."""
+        model = _build_flow_model()
+        i = _flow_dit_inputs()
+        out = model.dit_forward(i["noisy_action"], i["t"], **_flow_dit_kwargs(i))
+        assert out.shape == (BATCH, ACTION_LEN, ACTION_DIM)
+
+    def test_prefix_length_zeroes_leading_actions(self) -> None:
+        """A positive prefix_length forces the leading action tokens to zero."""
+        model = _build_flow_model()
+        i = _flow_dit_inputs()
+        kwargs = _flow_dit_kwargs(i)
+        kwargs["prefix_length"] = 2
+        out = model.dit_forward(i["noisy_action"], i["t"], **kwargs)
+        torch.testing.assert_close(out[:, :2], torch.zeros(BATCH, 2, ACTION_DIM))
+
+    @pytest.mark.parametrize(
+        "reference",
+        [torch.tensor([0.15467618, 0.23847848, 0.09961469, -0.11562672, 0.05315172])],
+    )
+    def test_reference(self, reference: torch.Tensor) -> None:
+        """A seeded dit_forward pins a slice of its velocity output."""
+        torch.manual_seed(0)
+        model = _build_flow_model()
+        i = _flow_dit_inputs()
+        out = model.dit_forward(i["noisy_action"], i["t"], **_flow_dit_kwargs(i))[0, 0, :5]
+        assert torch.allclose(out, reference, **TOL), out.tolist()
+
+
+class TestFlowGenerate:
+    """Tests for the Euler integration inference loop."""
+
+    @pytest.mark.parametrize(
+        "reference",
+        [torch.tensor([-1.96027327, -0.91721338, 0.38879472, -0.24504544, 0.8044914])],
+    )
+    def test_reference(self, reference: torch.Tensor) -> None:
+        """A seeded _flow_generate pins a slice of its Euler-integrated output."""
+        torch.manual_seed(0)
+        model = _build_flow_model()
+        i = _flow_dit_inputs()
+        x0 = torch.randn(BATCH, ACTION_LEN, ACTION_DIM)
+        out = model._flow_generate(x0, _flow_dit_kwargs(i))[0, 0, :5]
         assert torch.allclose(out, reference, **TOL), out.tolist()
