+++ library/tests/unit/policies/xr0/test_policy.py
@@ -24,15 +24,7 @@ from physicalai.export.backends import TorchExportParameters
 from physicalai.inference.data import InferenceFeatureType
 from physicalai.policies import get_physicalai_policy_class, get_policy
 from physicalai.policies.xr0 import XR0, XR0Config
-from physicalai.policies.xr0.export_openvino import (
-    export_add_deepstack_embeds,
-    export_build_additive_causal_mask,
-    export_fast_pos_embed_interpolate,
-    export_rot_pos_emb,
-    export_scatter_visual_embeds,
-    export_vision_attn_forward,
-    ov_friendly_rmsnorm_forward,
-)
+
 
 
 def _minimal_export_stats() -> dict[str, dict[str, Any]]:
@@ -225,202 +217,3 @@ class TestXR0Export:
         # Two cameras -> per-view names (no single-camera IMAGES collapse).
         assert names == {STATE, f"{IMAGES}.base", f"{IMAGES}.wrist_left", TASK}
 
-
-class TestExportPatchParity:
-    """Numerical parity of the export-friendly VLM ops against stock Qwen3-VL.
-
-    ``XR0Qwen3VL._ensure_export_patch`` swaps a handful of stock Qwen3-VL ops for
-    OpenVINO-friendly reimplementations (see the ``export_*`` module-level
-    functions). These tests check each replacement independently, comparing it to
-    the stock ``transformers`` op on small reference tensors. They deliberately
-    build tiny weight-free / small-module fixtures instead of loading the 4B model
-    so they stay fast and download-free; the export patch is numerically identical
-    to stock, so the outputs must match to floating-point tolerance.
-    """
-
-    def test_rot_pos_emb_matches_stock(self) -> None:
-        """``export_rot_pos_emb`` matches stock ``rot_pos_emb`` (weight-free)."""
-        from transformers.models.qwen3_vl.modeling_qwen3_vl import (
-            Qwen3VLVisionModel,
-            Qwen3VLVisionRotaryEmbedding,
-        )
-
-        torch.manual_seed(0)
-        merge_size = 2
-        # rotary emb is a deterministic ``inv_freq`` buffer -> no learned weights.
-        visual = types.SimpleNamespace(
-            spatial_merge_size=merge_size,
-            rotary_pos_emb=Qwen3VLVisionRotaryEmbedding(8),
-        )
-        # Two images (multi-frame on the second) to exercise the repeat path.
-        grid_thw = torch.tensor([[1, 4, 6], [2, 2, 8]])
-
-        stock = Qwen3VLVisionModel.rot_pos_emb(visual, grid_thw)
-        exported = export_rot_pos_emb(visual, grid_thw.tolist())
-
-        assert exported.shape == stock.shape
-        assert torch.equal(exported, stock)
-
-    def test_fast_pos_embed_interpolate_matches_stock(self) -> None:
-        """``export_fast_pos_embed_interpolate`` matches the stock interpolation."""
-        from transformers.models.qwen3_vl.modeling_qwen3_vl import Qwen3VLVisionModel
-
-        torch.manual_seed(0)
-        num_grid_per_side = 6
-        hidden = 16
-        merge_size = 2
-        pos_embed = torch.nn.Embedding(num_grid_per_side * num_grid_per_side, hidden)
-        visual = types.SimpleNamespace(
-            num_grid_per_side=num_grid_per_side,
-            pos_embed=pos_embed,
-            config=types.SimpleNamespace(spatial_merge_size=merge_size),
-        )
-        grid_thw = torch.tensor([[1, 4, 6], [1, 2, 8]])
-
-        stock = Qwen3VLVisionModel.fast_pos_embed_interpolate(visual, grid_thw)
-        exported = export_fast_pos_embed_interpolate(visual, grid_thw.tolist())
-
-        assert exported.shape == stock.shape
-        assert torch.allclose(exported, stock, atol=1e-6)
-
-    def test_vision_attn_forward_matches_stock(self) -> None:
-        """``export_vision_attn_forward`` matches stock attention (SDPA path)."""
-        from transformers.models.qwen3_vl.configuration_qwen3_vl import Qwen3VLVisionConfig
-        from transformers.models.qwen3_vl.modeling_qwen3_vl import Qwen3VLVisionAttention
-
-        torch.manual_seed(0)
-        config = Qwen3VLVisionConfig(hidden_size=32, num_heads=4)
-        config._attn_implementation = "sdpa"
-        attn = Qwen3VLVisionAttention(config).eval()
-        head_dim = config.hidden_size // config.num_heads
-
-        # Two attention windows of 6 and 4 tokens -> seq_len 10.
-        split_sizes = [6, 4]
-        seq_len = sum(split_sizes)
-        cu_seqlens = torch.tensor([0, 6, 10])
-        hidden_states = torch.randn(seq_len, config.hidden_size)
-        cos = torch.randn(seq_len, head_dim)
-        sin = torch.randn(seq_len, head_dim)
-        position_embeddings = (cos, sin)
-
-        with torch.no_grad():
-            stock = attn.forward(
-                hidden_states,
-                cu_seqlens,
-                position_embeddings=position_embeddings,
-            )
-            exported = export_vision_attn_forward(
-                attn,
-                split_sizes,
-                hidden_states,
-                position_embeddings,
-            )
-
-        assert exported.shape == stock.shape
-        assert torch.allclose(exported, stock, atol=1e-5)
-
-    def test_scatter_visual_embeds_matches_masked_scatter(self) -> None:
-        """``export_scatter_visual_embeds`` matches stock ``masked_scatter`` merge."""
-        torch.manual_seed(0)
-        seq_len, hidden, num_visual = 8, 4, 3
-        inputs_embeds = torch.randn(1, seq_len, hidden)
-        image_token_indices = torch.tensor([2, 4, 5])
-        image_embeds = torch.randn(num_visual, hidden)
-
-        # Stock merge: broadcast a boolean mask and ``masked_scatter``.
-        image_mask = torch.zeros(1, seq_len, dtype=torch.bool)
-        image_mask[0, image_token_indices] = True
-        image_mask = image_mask.unsqueeze(-1).expand_as(inputs_embeds)
-        stock = inputs_embeds.masked_scatter(image_mask, image_embeds)
-
-        exported = export_scatter_visual_embeds(inputs_embeds, image_token_indices, image_embeds)
-
-        assert exported.shape == stock.shape
-        assert torch.equal(exported, stock)
-
-    def test_add_deepstack_embeds_matches_stock(self) -> None:
-        """``export_add_deepstack_embeds`` matches stock ``_deepstack_process``."""
-        from transformers.models.qwen3_vl.modeling_qwen3_vl import Qwen3VLTextModel
-
-        torch.manual_seed(0)
-        seq_len, hidden, num_visual = 8, 4, 3
-        hidden_states = torch.randn(1, seq_len, hidden)
-        image_token_indices = torch.tensor([1, 3, 6])
-        visual_embeds = torch.randn(num_visual, hidden)
-
-        # Stock deepstack uses a boolean mask over the flattened (batch, seq) grid.
-        visual_pos_masks = torch.zeros(1, seq_len, dtype=torch.bool)
-        visual_pos_masks[0, image_token_indices] = True
-        stock = Qwen3VLTextModel._deepstack_process(
-            types.SimpleNamespace(),
-            hidden_states,
-            visual_pos_masks,
-            visual_embeds,
-        )
-
-        exported = export_add_deepstack_embeds(hidden_states, image_token_indices, visual_embeds)
-
-        assert exported.shape == stock.shape
-        assert torch.allclose(exported, stock, atol=1e-6)
-
-    @pytest.mark.parametrize("dtype", [torch.float32, torch.bfloat16])
-    @pytest.mark.parametrize(
-        "attention_mask",
-        [
-            [[1, 1, 1, 1, 1]],  # no padding -> pure causal
-            [[1, 1, 1, 0, 0]],  # right padding
-            [[1, 1, 1, 0, 0], [1, 1, 1, 1, 1]],  # batched, mixed padding
-        ],
-    )
-    def test_build_additive_causal_mask_matches_stock(
-        self,
-        attention_mask: list[list[int]],
-        dtype: torch.dtype,
-    ) -> None:
-        """``export_build_additive_causal_mask`` matches stock ``eager_mask``."""
-        from transformers.masking_utils import eager_mask
-
-        mask = torch.tensor(attention_mask, dtype=torch.long)
-        batch, seq_len = mask.shape
-
-        stock = eager_mask(
-            batch_size=batch,
-            q_length=seq_len,
-            kv_length=seq_len,
-            attention_mask=mask.to(torch.bool),
-            dtype=dtype,
-        )
-        exported = export_build_additive_causal_mask(mask, dtype)
-
-        assert exported.shape == (batch, 1, seq_len, seq_len)
-        assert exported.shape == stock.shape
-        assert exported.dtype == stock.dtype
-        assert torch.equal(exported, stock)
-
-    @pytest.mark.parametrize("dtype", [torch.float32, torch.bfloat16])
-    @pytest.mark.parametrize("shape", [(2, 16), (2, 5, 16), (1, 3, 4, 16)])
-    def test_ov_friendly_rmsnorm_matches_stock(
-        self,
-        shape: tuple[int, ...],
-        dtype: torch.dtype,
-    ) -> None:
-        """``ov_friendly_rmsnorm_forward`` matches stock ``Qwen3VLTextRMSNorm``."""
-        from transformers.models.qwen3_vl.modeling_qwen3_vl import Qwen3VLTextRMSNorm
-
-        torch.manual_seed(0)
-        hidden = shape[-1]
-        norm = Qwen3VLTextRMSNorm(hidden).eval()
-        with torch.no_grad():
-            # Randomize the weight so the weight-scaling path is exercised.
-            norm.weight.copy_(torch.randn(hidden))
-        x = torch.randn(*shape, dtype=dtype)
-
-        with torch.no_grad():
-            # Clone per call: the export forward reduces over ``dim() - 1`` in
-            # place on its float32 copy, which for a float32 input would alias x.
-            stock = norm(x.clone())
-            exported = ov_friendly_rmsnorm_forward(norm, x.clone())
-
-        assert exported.shape == stock.shape
-        assert exported.dtype == stock.dtype
-        assert torch.equal(exported, stock)
